Reorganize Queues
This moves the queues to rather than being local variables to being
members of the PodController struct. This means we don't need to
pass them around as much.
